Modernize signac-flow usage.
